Add deploy script and starter html file
I went with a simple deploy script for now. I'll likely want the
transfer to work via a GitHub Action in the future, but for now this
was a pretty simple way to get started (issues with ssh configuration on
the web hosting server aside).

Now that I now I can deploy and see things on the web server, I'm ready
to start on the initial minimal requirements.